Added a new class that helps with how the toolbar is created
